Add files via upload
Hi Franck, I have attempted to allow a choice of three and five rounds by creating new functions "quiz_main_3" & "quiz_main_5" in lines 65 & 101 respectively, and by calling them within the function "round_selection" in line 177. If the input is not equal to 3 or 5, the error is captured. However, the program just quits at this point. I see a lot of errors relating to variables not being defined, or "cannot determine type".  As you instructed, I tried commenting out the import random section  but it seemed to make little difference.